pyproject.toml
[tool.poetry]
name = "aizynthfinder"
version = "2.4.0"
description = "Retrosynthetic route finding using neural network guided Monte-Carlo tree search"
authors = ["Molecular AI group <samuel.genheden@astrazeneca.com>"]
license = "MIT"
include = ["aizynthfinder/data/*.yml", "aizynthfinder/data/templates/*"]
readme = "README.md"

[tool.pytest.ini_options]
mccabe-complexity = 9

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
apted = "^1.0.3"
deprecated = "^1.2.10"
grpcio = "^1.24.0"
ipywidgets = "^7.5.1"
jinja2 = "^2.11.0"
jupyter = "^1.0.0"
jupytext = "^1.3.3"
matplotlib = "^3.0.0"
more-itertools = "^8.2.0"
networkx = "^2.4"
pandas = "^1.0.0"
pillow = "^8.0.0"
pymongo = "^3.10.1"
rdchiral = "^1.0.0"
requests = "^2.23.0"
scikit-learn = "^0.24.0"
scipy = "^1.0"
tables = "^3.6.1"
tensorflow = "^2.1.0"
tensorflow-serving-api = "^2.1.0"
tqdm = "^4.42.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
invoke = "^1.5.0"
pytest = "^6.2.2"
pytest-black = "^0.3.12"
pytest-cov = "^2.11.0"
pytest-datadir = "^1.3.1"
pytest-mock = "^3.5.0"
pytest-mccabe = "^2.0.0"
sphinx = "^3.0.0"
mypy = "^0.800"

[tool.poetry.scripts]
aizynthapp = "aizynthfinder.interfaces.aizynthapp:main"
aizynthcli = "aizynthfinder.interfaces.aizynthcli:main"
aizynth_training = "aizynthfinder.training.training:main"
cat_aizynth_output = "aizynthfinder.tools.cat_output:main"
download_public_data = "aizynthfinder.tools.download_public_data:main"
make_false_products = "aizynthfinder.training.make_false_products:main"
preprocess_expansion = "aizynthfinder.training.preprocess_expansion:main"
preprocess_filter = "aizynthfinder.training.preprocess_filter:main"
preprocess_recommender = "aizynthfinder.training.preprocess_recommender:main"
preprocess_rollout = "aizynthfinder.training.preprocess_expansion:main"
smiles2stock = "aizynthfinder.tools.make_stock:main"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"